requirements.txt
colorama==0.4.6
rich==13.5.2
pyyaml==6.0.1
Flask==3.0.0
google-generativeai>=0.5.0  # For Gemini API 